fix: remove action entity references when deleting item
